+++ packages/signals_core/lib/src/extensions/future.dart
@@ -11,33 +11,6 @@ extension SignalFutureUtils<T> on Future<T> {
   /// final future = Future(() => 1);
   /// final signal = future.toSignal();
   /// ```
-  @Deprecated('use [toFutureSignal] instead')
-  FutureSignal<T> toSignal({
-    Duration? timeout,
-    String? debugLabel,
-    T? initialValue,
-    bool autoDispose = false,
-    bool lazy = true,
-    List<ReadonlySignal<dynamic>> dependencies = const [],
-  }) {
-    return toFutureSignal(
-      debugLabel: debugLabel,
-      initialValue: initialValue,
-      autoDispose: autoDispose,
-      lazy: lazy,
-      dependencies: dependencies,
-      timeout: timeout,
-    );
-  }
-
-  /// Convert an existing future to [FutureSignal]
-  ///
-  /// ```dart
-  /// import 'package:signals/signals.dart';
-  ///
-  /// final future = Future(() => 1);
-  /// final signal = future.toSignal();
-  /// ```
   FutureSignal<T> toFutureSignal({
     Duration? timeout,
     String? debugLabel,